feat: upgrade function hashPageRoute
