+++ src/styles/_sprinter-game.scss
@@ -386,3 +386,7 @@ $colorLevelSix: rgb(255, 128, 43);
   margin-top: 10px;
 }
 
+.sprint-game-close_span:hover {
+  transition: 0.3ms;
+  color: #6B53FF;
+}
