+++ src/views/renderPageGames.ts
@@ -2,17 +2,17 @@ function gamesView() {
   const viewGamesPage = `<section class="game-wrapper">
                           <div class="wrapper-games">
                             <div class="game-card sprint">
-                              <img class="game-img sprint" src="./assets/img/sprint.jpeg" alt="Image sprint game">
-                              <h2 class="sprint">Sprint</h2>
-                              <div class="wrapper-img sprint">
-                                <img class="icon-timer sprint" src="./assets/img/timer.svg"></img>
+                              <img class="game-img" src="./assets/img/sprint.jpeg" alt="Image sprint game">
+                              <h2>Sprint</h2>
+                              <div class="wrapper-img">
+                                <img class="icon-timer" src="./assets/img/timer.svg"></img>
                               </div>
                             </div>
                             <div class="game-card audio" id="game-audio">
-                              <img class="game-img audio" src="./assets/img/audio.jpeg" alt="Image audio game">
-                              <h2 class="audio">Audio game</h2>
-                              <div class="wrapper-img audio">
-                                <img class="icon-audio audio" src="./assets/img/sound.svg"></img>
+                              <img class="game-img" src="./assets/img/audio.jpeg" alt="Image audio game">
+                              <h2>Audio game</h2>
+                              <div class="wrapper-img">
+                                <img class="icon-audio" src="./assets/img/sound.svg"></img>
                               </div>
                             </div>
                           </div>
